Tests/LibWeb/Text/expected/css/filter-drop-shadow-color-accumulation.txt

0: drop-shadow(rgb(0, 0, 255) 4px 6px 0px)
500: drop-shadow(rgb(128, 0, 255) 5px 7px 0px)
1000: drop-shadow(rgb(255, 0, 255) 6px 8px 0px)
